move fpath function to utils
